chore: Move the lsp server to src/lsp module
So it can be used by the emscripten bindings.
